Update the reviewer for closed PRs version 1
